% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminants.R
\name{contaminant_AC}
\alias{contaminant_AC}
\title{Identify contaminants in a given vector of protein ACs.}
\usage{
contaminant_AC(protein_ac)
}
\arguments{
\item{protein_ac}{A vector of protein AC's.}
}
\value{
A vector indicating if it is a contaminant.
}
\description{
Gives back a boolean for every given protein AC, indicating if it is a
contaminant or not.
}
\examples{
contaminant_AC(c("A0AVF1", "Q9NZT1", "P02786"))
}
