% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminants.R
\name{contaminant_gene}
\alias{contaminant_gene}
\title{Identify contaminants by gene names}
\usage{
contaminant_gene(gene_names)
}
\arguments{
\item{gene_name}{A vector of gene names.}
}
\value{
A vector indicating if it is a contaminant.
}
\description{
Gives back a boolean for every given gene name, indicating if it is a
contaminant or not.
}
\examples{
contaminant_gene(c("TFRC", "TGM1", "MTFR1"))
}
