% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminants.R
\name{contaminant_term}
\alias{contaminant_term}
\title{Identify contaminants by terms}
\usage{
contaminant_term(protein_terms)
}
\arguments{
\item{protein_terms}{A vector of terms.}
}
\value{
A vector indicating if it is a contaminant.
}
\description{
Gives back a boolean for every given term, indicating if it is a
contaminant or not.
}
\examples{
contaminant_term(c("blibla", "blibla keratin blabla", "Keratinblibla", "kerami mouou"))
}
