% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminants_MQ.R
\name{contaminants_MQ}
\alias{contaminants_MQ}
\title{Find contaminants}
\usage{
contaminants_MQ(proteinGroups_df)
}
\arguments{
\item{proteinGroups_df}{A data.frame from proteinGroups.txt}
}
\description{
Find all contaminants in a data.frame from MaxQuant proteinGroups.txt.
}
