% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_contaminants.R
\name{kuravsky_transform_long}
\alias{kuravsky_transform_long}
\title{kuravsky_transform_long}
\usage{
kuravsky_transform_long(kuravsky_path)
}
\arguments{
\item{kuravsky_path}{Path to kuravsky_more_contamin.txt.}
}
\description{
There are several protein IDs and Gene.names per line in a MaxQuant
proteinGroups.txt file. We create a new line for every single protein Id.
}
\examples{
kuravsky_path <- "/Users/admin/Work/PAF/projects/SliceSILAC/latest/data/Kuravsky_7956_7999/kuravsky_more_contamin.txt"
contaminants_kuravsky <- kuravsky_transform_long(kuravsky_path)
devtools::use_data(contaminants_kuravsky)
}
